Fix a number of compiler warnings.
The main one is renaming discoverPlugins(QString) to discoverPlugin(QString),
to avoid the clash with discoverPlugins(), triggering -Woverloaded-virtual


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14282 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/sound/audiostream/WavFileWriteStream.cpp
@@ -72,7 +72,7 @@ WavFileWriteStream::putInterleavedFrames(size_t count, float *frames)
 
     sf_count_t written = sf_writef_float(m_file, frames, count);
 
-    return (written == count);
+    return (size_t(written) == count);
 }
 
 }
